change int to double

+++ ai/searchnode.cpp
@@ -392,7 +392,7 @@ double SearchNode::evaluate()
 	//rate4: distance to different positions in different cases
 	SceneState s=*state;
 	analyzer temp_analyzer(&s);
-	int rate1,rate2,rate3;
+	double rate1,rate2,rate3;
 	double food_density=temp_analyzer.best_way()[0].weight+temp_analyzer.best_way()[1].weight+temp_analyzer.best_way()[2].weight;
 	int extra_ability=0;
 	for (int i=SkillType(LONG_ATTACK);i<=SkillType(HEALTH_UP);i++)
@@ -432,8 +432,8 @@ double SearchNode::evaluate()
 		//evaluation
 		if (original_hp_rate>1) //Add enemy's ability points into consideration!
 		{
-			int rate4;
-			int rate5,rate6;//rate5: shield status, rate6: dash status
+			double rate4;
+			double rate5,rate6;//rate5: shield status, rate6: dash status
 			return original_hp_rate*state->status->objects[0].health+
 				rate1*(state->status->objects[0].ability-3/* 3 should be ability points of enemy, do we have that?*/)+
 				rate4*distance+
@@ -442,8 +442,8 @@ double SearchNode::evaluate()
 		}
 		else
 		{
-			int rate4;
-			int rate5,rate6;//rate5: shield status, rate6: dash status
+			double rate4;
+			double rate5,rate6;//rate5: shield status, rate6: dash status
 			return original_hp_rate*state->status->objects[0].health+
 				rate4*distance+
 				rate5*(state->status->objects[0].shield_time>0)+
@@ -452,7 +452,7 @@ double SearchNode::evaluate()
 	}
 	else if (enemy_in_sight==0&&boss_in_sight==0)									//Regular mode
 	{
-		int rate4;
+		double rate4;
 		return state->status->objects[0].health+
 			rate1*extra_ability+
 			rate2*food_density+
@@ -461,7 +461,7 @@ double SearchNode::evaluate()
 	else if(enemy_in_sight==0&&boss_in_sight==1)//Only boss in sight
 	{
 		//initializing data
-		int rate4;
+		double rate4;
 		int possible_extra_hp,hp_price=pow(2,state->status->objects[0].skill_level[SkillType(HEALTH_UP)]);
 		int temp_ability=state->status->objects[0].ability;
 		double kill_rate=pow(1.2,3),death_rate=1/kill_rate;
@@ -521,12 +521,12 @@ double SearchNode::evaluate()
 			gameover_state=1;
 
 		//evaluating
-		int rate_boss,rate_enemy,rate_boss_distance;
+		double rate_boss,rate_enemy,rate_boss_distance;
 		int enemy_evaluation_point;
 		if (Enemy_original_hp_rate>1) //Add enemy's ability points into consideration!
 		{
-			int rate4;
-			int rate5,rate6;//rate5: shield status, rate6: dash status
+			double rate4;
+			double rate5,rate6;//rate5: shield status, rate6: dash status
 			enemy_evaluation_point=	
 				Enemy_original_hp_rate*state->status->objects[0].health+
 				rate1*(state->status->objects[0].ability-3/* 3 should be ability points of enemy, do we have that?*/)+
@@ -536,8 +536,8 @@ double SearchNode::evaluate()
 		}
 		else
 		{
-			int rate4;
-			int rate5,rate6;//rate5: shield status, rate6: dash status
+			double rate4;
+			double rate5,rate6;//rate5: shield status, rate6: dash status
 			enemy_evaluation_point=
 				Enemy_original_hp_rate*state->status->objects[0].health+
 				rate4*Enemy_distance+
